refactor: replace some files and change imports
