+++ src/Components/App/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
-import './App.scss';
-import Header from './../Header/Header';
-import Main from '../Main/Main';
+import Header from '../Header';
+import Main from '../Main';
 
 const App = () => {
     return(
