+++ src/Components/Description/Description.js
@@ -1,28 +1,40 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import './Description.scss';
-import Initial from './Initial';
 import Player from '../Player/Player';
 import data from '../../data/data';
 
-function Description({ gameId, currentTheme }) {
+const Description = ({ gameId, currentTheme }) => {
     let pickedGame = data[currentTheme][gameId];
     return(
     <section className = "description">
-        {!pickedGame && <Initial />}
-        {pickedGame && 
-        <>
-        <img src = {pickedGame.image} alt = 'game poster' />
-        <div className = 'game-info'>
-            <h3 className = 'game-name'>{pickedGame.name}</h3>
-            <Player
-                audio = {pickedGame.audio}
-            />
-            <div className = 'game-description'>
-                {pickedGame.description}
-            </div>
-        </div>
-        </>}
+        {
+            !pickedGame
+            &&
+            (
+                <div className = 'initial'>
+                    <p>Послушайте плеер.</p>
+                    <p>Выберите игру из списка.</p>
+                </div>
+            )
+        }
+        {
+            pickedGame
+            &&
+            (
+                <>
+                    <img src = {pickedGame.image} alt = 'game poster' />
+                    <div className = 'game-info'>
+                        <h3 className = 'game-name'>{pickedGame.name}</h3>
+                        <Player
+                            audio = {pickedGame.audio}
+                        />
+                        <div className = 'game-description'>
+                            {pickedGame.description}
+                        </div>
+                    </div>
+                </>
+            )
+        }
     </section>
     );
 }
