--- src/Components/Description/Initial.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function Initial() {
-    return(
-        <div className = 'initial'>
-            <p>Послушайте плеер.</p>
-            <p>Выберите игру из списка.</p>
-        </div>
-    );
-}
-
-export default Initial;
