+++ src/Components/Description/index.js
@@ -1,0 +1,3 @@
+import Description from './Description';
+
+export default Description;
