+++ src/Components/Header/Header.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import './Header.scss';
 import TitleList from './TitleList';
 
 const Header = () => {
