+++ src/Components/NextButton/NextButton.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import './NextButton.scss';
 import {
     onChangeTitle,
     onRestartGame,
