+++ src/Components/NextButton/index.js
@@ -1,0 +1,3 @@
+import NextButton from './NextButton';
+
+export default NextButton;
