+++ src/Components/Player/index.js
@@ -1,0 +1,3 @@
+import Player from './Player';
+
+export default Player;
