+++ src/Components/Quiz/index.js
@@ -1,0 +1,3 @@
+import Quiz from './Quiz';
+
+export default Quiz;
