+++ src/Components/Results/index.js
@@ -1,0 +1,3 @@
+import Results from './Results';
+
+export default Results;
