+++ src/styles/_app.scss
@@ -1,11 +1,11 @@
 @font-face {
     font-family: 'Pricedown';
-    src: url('./assets/Pricedown.ttf') format('truetype');
+    src: url('../assets/Pricedown.ttf') format('truetype');
 }
 
 @font-face {
     font-family: 'Roboto';
-    src: url('./assets/Roboto-Regular.ttf') format('truetype');
+    src: url('../assets/Roboto-Regular.ttf') format('truetype');
 }
 
 * {
