+++ src/styles/_description.scss
