+++ src/styles/_quiz.scss
