+++ src/styles/_results.scss
