+++ src/styles/index.scss
@@ -1,0 +1,8 @@
+@import "app";
+@import "description";
+@import "header";
+@import "main";
+@import "nextButton";
+@import "question";
+@import "quiz";
+@import "results";
